Separated JNA callback interfaces to DebugCallback.java and ErrorCallback.java.
